Mark WDT section ESP8266-specific
    PUBLISHED_FROM=6181267507a1f68a4bb02f82814fb4f51c8059d6

+++ src/v7.c
@@ -4403,7 +4403,7 @@ ON_FLASH int c_vsnprintf(char *buf, size_t buf_size, const char *fmt,
 
   /* Zero-terminate the result */
   if (buf_size > 0) {
-    buf[i < (int) buf_size ? i : buf_size - 1] = '\0';
+    buf[i < (int) buf_size ? i : (int) buf_size - 1] = '\0';
   }
 
   return i;
